v0.5.0

+++ app/controllers/questions_controller.rb
@@ -1,21 +1,23 @@
+# frozen_string_literal: true
+
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
   def index
-    @questions=Question.all
+    @questions = Question.all
   end
 
   def show
+    @answer ||= question.answers.new
   end
 
-  def new
-  end
+  def new; end
 
-  def edit
-  end
+  def edit; end
 
   def create
-    @question = Question.new(question_params)
+    @question = current_user.questions.new(question_params)
     if @question.save
-    redirect_to @question
+      redirect_to @question, notice: 'Your question sucessfully created.'
     else
       render :new
     end
@@ -30,19 +32,23 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user&.author?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Question delited sucessfully'
+    else
+      redirect_to questions_path, alert: 'You are not owner of this question'
+    end
   end
 
   private
 
   def question
-    @question ||=params[:id] ? Question.find(params[:id]) : Question.new
+    @question ||= params[:id] ? Question.find(params[:id]) : Question.new
   end
 
   helper_method :question
 
   def question_params
-    params.require(:question).permit(:title,:body)
+    params.require(:question).permit(:title, :body)
   end
 end
